headers: Add data_race(expr)
This define is just a hint for the compiler that it should ignore the
data race.
It was added to mainline Linux kernel in commit c48981eeb0d5
("include/linux/compiler.h: Introduce data_race(expr) macro") with
kernel 5.8 to provide a hint to the compiler and the reviewer that there
is an intentional data race.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>