backports: increase cfg80211 wiphy padding
The wiphy structure has grown over the 2048 bytes that we originally
used in the wiphy padding, it's 2080 bytes long as of v5.11.  Increase
the padding to an arbitrary higher round number, 2304.  We will end up
using 2 pages now, so we could go as high as a page's length, but keep
it lower so that when we grow over 2304 again we notice and can reasses
the issue.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>